refactor: separate auth from App.jsx and split each step into pages
